fixes for dtype problems

+++ tests/test_blend.py
@@ -4,7 +4,7 @@ from numpy.testing import assert_almost_equal, assert_array_equal
 import scarlet
 
 
-def init_data(shape, coords, amplitudes=None, convolve=True):
+def init_data(shape, coords, amplitudes=None, convolve=True, dtype=np.float32):
     import scipy.signal
 
     B, Ny, Nx = shape
@@ -15,9 +15,9 @@ def init_data(shape, coords, amplitudes=None, convolve=True):
     assert K == len(amplitudes)
 
     _seds = [
-        np.arange(B, dtype=float),
-        np.arange(B, dtype=float)[::-1],
-        np.ones((B,), dtype=float)
+        np.arange(B, dtype=dtype),
+        np.arange(B, dtype=dtype)[::-1],
+        np.ones((B,), dtype=dtype)
     ]
     seds = np.array([_seds[n % 3]*amplitudes[n] for n in range(K)])
 
@@ -35,15 +35,15 @@ def init_data(shape, coords, amplitudes=None, convolve=True):
         target_psf /= target_psf.sum()
 
         psfs = np.array([scarlet.psf.generate_psf_image(scarlet.psf.gaussian, psf_shape, psf_center,
-                                                        amplitude=1, sigma=1+.2*b) for b in range(B)])
+                                                        amplitude=1, sigma=1+.2*b) for b in range(B)], dtype=dtype)
         # Convolve the image with the psf in each channel
         # Use scipy.signal.convolve without using FFTs as a sanity check
         images = np.array([scipy.signal.convolve(img, psf, method="direct", mode="same")
-                           for img, psf in zip(images, psfs)])
+                           for img, psf in zip(images, psfs)], dtype=dtype)
         # Convolve the true morphology with the target PSF,
         # also using scipy.signal.convolve as a sanity check
         morphs = np.array([scipy.signal.convolve(m, target_psf, method="direct", mode="same")
-                           for m in morphs])
+                           for m in morphs], dtype=dtype)
         morphs /= morphs.max()
         psfs /= psfs.sum(axis=(1,2))[:,None,None]
 
@@ -56,10 +56,10 @@ class TestBlend(object):
     def test_init(self):
         shape = (6, 31, 55)
         coords = [(20, 10), (10, 30), (17, 42)]
-        target_psf, psfs, images, channels, seds, morphs = init_data(shape, coords, [3, 2, 1])
+        target_psf, psfs, images, channels, seds, morphs = init_data(shape, coords, [3, 2, 1], dtype=np.float64)
 
         # Test vanilla init
-        frame = scarlet.Frame(images.shape)
+        frame = scarlet.Frame(images.shape, dtype=np.float64)
         observation = scarlet.Observation(images).match(frame)
         sources = [scarlet.PointSource(frame, coord, observation) for coord in coords]
         blend = scarlet.Blend(sources, observation)
@@ -71,7 +71,7 @@ class TestBlend(object):
         assert blend.frame.psfs == frame.psfs
 
         # Test init with psfs
-        frame = scarlet.Frame(images.shape, psfs=target_psf[None])
+        frame = scarlet.Frame(images.shape, psfs=target_psf[None], dtype=np.float64)
         observation = scarlet.Observation(images, psfs=psfs).match(frame)
         sources = [scarlet.PointSource(frame, coord, observation) for coord in coords]
 
@@ -93,11 +93,11 @@ class TestBlend(object):
         shape = (6, 31, 55)
         coords = [(20, 10), (10, 30), (17, 42)]
         amplitudes = [3, 2, 1]
-        target_psf, psfs, images, channels, seds, morphs = init_data(shape, coords, amplitudes)
+        target_psf, psfs, images, channels, seds, morphs = init_data(shape, coords, amplitudes, dtype=np.float64)
         B, Ny, Nx = shape
         K = len(coords)
 
-        frame = scarlet.Frame(images.shape, psfs=target_psf[None])
+        frame = scarlet.Frame(images.shape, psfs=target_psf[None], dtype=np.float64)
         observation = scarlet.Observation(images, psfs=psfs).match(frame)
         sources = [scarlet.PointSource(frame, coord, observation) for coord in coords]
         blend = scarlet.Blend(sources, observation)
@@ -116,10 +116,10 @@ class TestBlend(object):
         shape = (6, 31, 55)
         coords = [(20, 10), (10, 30), (17, 42)]
         amplitudes = [3, 2, 1]
-        target_psf, psfs, images, channels, seds, morphs = init_data(shape, coords, amplitudes)
+        target_psf, psfs, images, channels, seds, morphs = init_data(shape, coords, amplitudes, dtype=np.float64)
         B, Ny, Nx = shape
 
-        frame = scarlet.Frame(images.shape, psfs=target_psf[None])
+        frame = scarlet.Frame(images.shape, psfs=target_psf[None], dtype=np.float64)
         observation = scarlet.Observation(images, psfs=psfs).match(frame)
         bg_rms = np.ones((B,))
         sources = [scarlet.ExtendedSource(frame, coord, observation, bg_rms) for coord in coords]
